Move GameLoader to Prototype.Networking.Startup
