+++ Assets/Prototype/Networking/Startup.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3cd9398689004284ba509ef98e5a2547
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
